Fixed compiler warnings

+++ libmorton/include/morton2D.h
@@ -95,7 +95,7 @@ template<typename morton, typename coord>
 inline morton morton2D_SplitBy2Bits(const coord a) {
 	const morton* masks = (sizeof(morton) <= 4) ? reinterpret_cast<const morton*>(magicbit2D_masks32) : reinterpret_cast<const morton*>(magicbit2D_masks64);
 	morton x = a;
-	if (sizeof(morton) > 4) { x = (x | x << 32) & masks[0]; }
+	if (sizeof(morton) > 4) { x = (x | (uint_fast64_t) x << 32) & masks[0]; }
 	x = (x | x << 16) & masks[1];
 	x = (x | x << 8) & masks[2];
 	x = (x | x << 4) & masks[3];
@@ -114,7 +114,7 @@ inline morton m2D_e_magicbits(const coord x, const coord y) {
 template<typename morton, typename coord>
 inline morton m2D_e_for(const coord x, const coord y){
 	morton answer = 0;
-	unsigned int checkbits = floor(sizeof(morton) * 4.0f);
+	unsigned int checkbits = (unsigned int) floor(sizeof(morton) * 4.0f);
 	for (unsigned int i = 0; i <= checkbits; ++i) {
 		morton mshifted = static_cast<morton>(0x1) << i; // Here we need to cast 0x1 to 64bits, otherwise there is a bug when morton code is larger than 32 bits
 		unsigned int shift = 2 * i;
